requirements.txt
langchain>=0.1.0
# langchain-google-genai>=0.0.11
langchain-ibm>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
mysql-connector-python>=8.0.0
pymysql>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0